Improve the backup and restore commands and implement the info command

+++ lib/bak-shell/commands/restore.rb
@@ -19,12 +19,16 @@ module BakShell
         raise InvalidBackupError.new("No backup found for file or directory: #{target}") if backup.nil?
 
         backup_dir = File.join(BakShell::BACKUP_DIR, backup.id)
-        base_target = File.basename(target)
-        latest_backup_dir = File.join(backup_dir, base_target)
+        latest_backup = Dir.glob(File.join(backup_dir, "*.bak-latest-*"))
 
-        temp_bak = "target.bak-#{Time.now.to_f}"
+        raise "No backup marked as latest was found" if latest_backup.empty?
+        raise "More than one backup marked as latest was found" if latest_backup.count > 1
+
+        latest_backup = latest_backup.first
+
+        temp_bak = "#{target}.bak-#{Time.now.to_f}"
         FileUtils.mv(target, temp_bak)
-        FileUtils.cp_r(latest_backup_dir, File.dirname(target))
+        FileUtils.cp_r(latest_backup, File.join(File.dirname(target), File.basename(latest_backup).sub(/\.bak-latest-.*\Z/, "")))
         FileUtils.rm_r(temp_bak, force: true)
 
         puts "Backup restored!".color(:green)
